Converted helpers into modules for better testing and due to convention. Modified controllers and views to accomadate this change. Configuration changes for modules
