+++ config/application.rb
@@ -22,5 +22,6 @@ module CtfRegistration
 
     # Do not swallow errors in after_commit/after_rollback callbacks.
     config.active_record.raise_in_transactional_callbacks = true
+    config.autoload_paths += %W(#{config.root}/lib)
   end
 end
