+++ config/initializers/application_module.rb
@@ -1,0 +1,1 @@
+require 'application_module.rb'
