+++ config/initializers/user_module.rb
@@ -1,0 +1,1 @@
+require 'user_module.rb'
